`exhaustive_cases` support for switch pattern cases
