+++ lib/src/rules/exhaustive_cases.dart
@@ -118,22 +118,28 @@ class _Visitor extends SimpleAstVisitor {
 
       var enumConstants = enumDescription.enumConstants;
       for (var member in statement.members) {
-        if (member is SwitchCase) {
-          var expression = member.expression.unParenthesized;
-          if (expression is Identifier) {
-            var element = expression.staticElement;
-            if (element is PropertyAccessorElement) {
-              enumConstants.remove(element.variable.computeConstantValue());
-            } else if (element is VariableElement) {
-              enumConstants.remove(element.computeConstantValue());
-            }
-          } else if (expression is PropertyAccess) {
-            var element = expression.propertyName.staticElement;
-            if (element is PropertyAccessorElement) {
-              enumConstants.remove(element.variable.computeConstantValue());
-            } else if (element is VariableElement) {
-              enumConstants.remove(element.computeConstantValue());
-            }
+        Expression? expression;
+        if (member is SwitchPatternCase) {
+          var pattern = member.guardedPattern.pattern.unParenthesized;
+          if (pattern is ConstantPattern) {
+            expression = pattern.expression.unParenthesized;
+          }
+        } else if (member is SwitchCase) {
+          expression = member.expression.unParenthesized;
+        }
+        if (expression is Identifier) {
+          var element = expression.staticElement;
+          if (element is PropertyAccessorElement) {
+            enumConstants.remove(element.variable.computeConstantValue());
+          } else if (element is VariableElement) {
+            enumConstants.remove(element.computeConstantValue());
+          }
+        } else if (expression is PropertyAccess) {
+          var element = expression.propertyName.staticElement;
+          if (element is PropertyAccessorElement) {
+            enumConstants.remove(element.variable.computeConstantValue());
+          } else if (element is VariableElement) {
+            enumConstants.remove(element.computeConstantValue());
           }
         }
         if (member is SwitchDefault) {
